bug fix in disk-failure-handling - copy mac address from parent vm
to fix data disk loss, we create an empty VM and attach system disks from parent vm.
If the mac address for the new vm is a new one, then linux probably will rename eth0
to eth1 according to udev rules, and thus our network initialization script will fail.
In this fix, we copied the mac address of parent vm to prevent the eth rename issue.