implement the inner host target selection algorithm
we improve the vm clone process by encouraging more clones inside the same host. The metric Scarcity as we
named, is to measure the urgency a host needs a clone. The scarcity basically evaluates how many rounds of
clone process is required to complete the cloning process for all nodes on that host.

The targetSelector is implemented as Strategy pattern in FastCloneService, i.e., you can inject different
selection algorithms based on your needs.